refactor(tools): move all forked tools (click-path, click-default-group and imageio-ffmpeg:io) to a 'tools/forked' dedicated directory
