+++ pyproject.toml
@@ -45,8 +45,7 @@ profile = "black"
 
 [tool.mypy]
 exclude = '''(?x)(
-    ^vhcalc/tools/click_path\.py$               # file named "vhcalc/tools/click_path.py"
-    | ^vhcalc/tools/imageio_ffmpeg_io\.py$      # file named "vhcalc/tools/imageio_ffmpeg_io.py"
+    ^vhcalc/tools/forked               # all files named in "vhcalc/tools/forked/" directory
 )'''
 files = [
     "vhcalc",
@@ -118,8 +117,7 @@ norecursedirs = [
         '*/virtualenv/*',
         '*/virtualenvs/*',
         '*/tests/*',
-        '*/tools/click_default_group.py',
-        '*/tools/imageio_ffmpeg_io.py'
+        '*/tools/forked/*'
     ]
     [tool.coverage.run]
     relative_files = true
