+++ setup.cfg
@@ -17,4 +17,4 @@ exclude =
 max-line-length = 88
 per-file-ignores =
     # https://black.readthedocs.io/en/stable/guides/using_black_with_other_tools.html#e701-e704
-    vhcalc/tools/click_default_group.py:W503,A001,E704
+    vhcalc/tools/forked/click_default_group.py:W503,A001,E704
