+++ vhcalc/services/__init__.py
@@ -1,11 +1,9 @@
 from .imghashes import (
-    compute_imghash_from_media,
     compute_imghash_from_media_from_binary_stream,
     export_imghash_from_media,
 )
 
 __all__ = [
     "export_imghash_from_media",
-    "compute_imghash_from_media",
     "compute_imghash_from_media_from_binary_stream",
 ]
